Fix wheelEvent parameter type
(cherry picked from commit 5bcfa2692b757ab03adb11ede6680f40407eb4dd)

+++ source/app/ui/graphcommoninteractor.cpp
@@ -433,7 +433,7 @@ NodeId GraphCommonInteractor::nodeIdNearPosition(const QPoint& localPosition) co
                                            frustum, ray.origin());
 }
 
-void GraphCommonInteractor::wheelEvent(const QPoint& pos, float angle)
+void GraphCommonInteractor::wheelEvent(const QPoint& pos, int angle)
 {
     if(_graphRenderer->transition().active())
         return;
